refactor: move EventPublisher to models layer to eliminate importlib hack in event_helpers
Moves EventPublisher/publish/subscribe from domain/publisher.py to
models/event_bus.py. Since DomainEvent is already in models, the
event bus has no domain-layer dependencies.

- domain/publisher.py becomes a re-export stub (backward compat)
- services/event_helpers.py uses clean static import instead of importlib
- No behavior change; same singleton EventPublisher semantics
